ovpn-dco: fix ovpn iface reference counting
When the user request the deletion of the ovpn interface,
we must ensure that the net_device object will stay around until
the last user has dropped its reference to it.

In our case, the peer object has a pointer to the ovpn priv object
therefore we must ensure that it also holds a reference that is
dropped when the peer is released.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>