ovpn-dco: fix ref imbalance along RX path in case of no crypto context
When no crypto context is allocated for a certain peer and we receive a
packet from that peer, ovpn_recv_crypto() will abort and exit without
attempting any decryption.

However, in case of abort the reference to the peer object is not
dropped, thus causing an imbalance.

Release peer reference in case of early exit and fix refcounting.

Reported-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>