ovpn-dco: let userspace inject the peer-ID of the remote host
When connecting to a server, the peer-id to use on the data channel is
communicated via push response to userspace. At this point userspace has
to communicate the received value to ovpn-dco so that the latter can use
it when constructing DATA packets.

Add the remote_peer_id as attribute of the SET_KEYS netlink command, so
that ovpn-dco can then store and it use it later on.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>