ovpn-dco: refactor crypto context and introduce concept of slot
The crypto_context_pair was not extremely useful, except that it allowed
the whole key pair to be handled as one object.

To increase flexibility, and simplify the code, the context_pair struct
has been removed and the context has been renamed to "key_slot".

The NEW_KEY netlink command (formely SET_KEY) allows the user to operate
on single key_slots, so to ensure that all the compl;exity is hidden in
userspace.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>