ovpn-dco: avoid headroom computation along TX path
The skb headroom for outbound packets is properly pre-allocated by the
netdevice itself thanks to the needed_headroom member filled upon device
initialization.

For this reason it is not needed to recompute the headroom everytime, but
it is enough to re-use the same value specified upon initialization.

This way the TX path can be slightly simplified and some operations can
be avoided.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>