ovpn-dco: refactor TX/RX path to use additional (interruptible) thread
Instead of performing the entire packet processing in one interrupt, it
is better (and allows more flexibility) to pass the packet to a second
(interruptible) thread that can perform all subsequent operation.

Passing packets is performed by using a ptr_ring and the second thread
is scheduled with a simple worker on a workqueue.

This is a ismple restructuring, so no multicore trick has been yet
applied.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>