ovpn-dco: greatly simplify encryption/decryption code
Now that the encryption/decryption code is running in its own kernel
thread (not in a packet related softirq anymore) it is possible to make
it sleep.

This possibility allows us to re-implement the crypto code in a way that
does not requires chains of callbacks to handle the asynchronous case
(i.e. crypto_aead_en/decrypt() returning -EINPROGRESS).

Get rid of callbacks and simplify code.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>