ovpn-dco: move mutex from peer to crypto_state struct
Mutex in struct peer was used to protect the pointer to
crypto_state.
However that pointer is now a static object and the mutex is used to
protect the keyslots inside it.

For this reason it makes sense to move the mutex inside the
crypto_state struct, next to the pointers it protects.

Signed-off-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>