ovpn-dco: remove useless ptr_ring cleanup function
A peer is released only when all references are dropped, which
translates in it not being selectable anymore for enqueing SKBs.

At the same time, if any SKB was previously queued, the worker would
still be running to consume them.

Since the worker holds a reference to the peer, we can get to its
release stage only if there are no more packets to process in the
ring.

For this reason we don't need any ptr_ring cleanup function.

We, however, WARN_ON releasing a peer that still has some data in the
rings, so that we can catch possible bugs.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>