ovpn-dco: implement keepalive with interval and expiry timeout
The OpenVPN protocol supports a keepaive mechanism that consists in
sending a VPN ping packet to the other peer every (configured) interval.

At the same time, a peer can be configured to purge a peer, if no data
is received for a certain amount of time. The ping ensures that this
timeout is never reached in case of no data.

Both the ping interval and the expiry timeout can now be configured via
netlink CMD_SET_PEER.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>