ovpn-dco: avoid race condition between scheduled work and RCU callback
When deleting a peer we have no guarantee that the RCU Callback will
always executed after the scheduled work, however, the code makes this
assumption.

To make sure we don't enter a race condition, schedule the RCU callback
direcly from the scheduled work. This way we now that the latter can't
be executed before the former.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>