ovpn-dco: implement TCP transport support
With this change ovpn-dco is now able to handle a TCP socket
at the transport layer to connect to another peer.

The socket is expected to be delivered to ovpn-dco via netlink
in ESTABLISHED state. At that point ovpn-dco will take over and
will start sending/receiving packets.

The data format is the same as the OpenVPN userspace implementation,
where each packet is prefixed with its actual size and then sent over
the stream. The receiver side will read the 2 bytes size, extract the
rest of the packet from the stream and convert it back to an skb.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>