ovpn-dco: extend ovpn-cli with TCP transport layer support
ovpn-cli is now able to create a TCP socket and pass it down to
ovpn-dco.

Since a TCP socket needs tobe created between 2 asymmetric peers,
ovpn-cli is extended with 2 new commands:
* listen <params>
* connect <params>

Once the connection is established, ovpn-cli will start the VPN and pass
the socket to kernelspace. This operation will automatically add the
peer using the endpoint the connection was received from.

After that the user has to still install the keys using the new_key
command.

At the same time update the netns-test.sh script to run in TCP mod when
requested.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>