ovpn-dco: fix crash on rmmod when peer creation was aborted
During peer creation in ovpn_peer_new(), ovpn-dco configures and starts
NAPI. If then creation is aborted because of error, peer memory is
freed.

When calling rmmod, kernel destroys net device and tries to use NAPI
context which was associated with it, which causes crash since it was
already freed as part of peer memory.

Fix by disabling NAPI and removing NAPI context from net device.

Signed-off-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>