ovpn-dco: prevent API calls if VPN has not yet been started
As of now there is no way to detect whether a VPN has been started or
not and therefore a user may send any command (i.e. NEW_PEER) even when
the VPN hasn't yet started.

Prevent this by using the "mode" as flag. When mode is UNDEF, the VPN is
considered not-started and therefore any netlink command will be
rejected, except for START_VPN.

When mode is anything else other netlink commands are accepted too.

This patch implicitly fixes all those bugs that are triggered by calling
netlink commands that expected various fileds to be already initialized,
like the sock object.

Reported-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>