ovpn-dco: prepend TCP payload with correct size
As required by OpenVPN protocol, after encryption we prepend the payload data
with its size. This way the receiver is able to extract the exact
datagram from the TCP stream.

Since we use GSO and therefore skb could be a list, while traversing it we
should use "curr" to access current each list item.
We erroneously use "skb", which means that we prepend all payloads in the
list with the size of the first item. This breaks the TCP connection.

Fix it by prepending the appropriate "curr->len" to the payload.

Signed-off-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>