ovpn-dco: avoid potential out of bound access in aead_decrypt()
The 'nfrags' returned by skb_cow_head() should be checked against the
'sg' array size decremented by 2, since in sg_table_init() we pass
'nfrags + 2' as table size. Therefore the check has to be fixed.

However, there is no need for the additional 2 elements in the array
because space for auth-tag and additional data is already allocated
as they are both part of the skb passed to skb_cow_data().

Fix the out of bound check by changing the size of the array and the
related initialization.

Addresses-Coverity: ("Out-of-bounds read")
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>