ovpn-dco: make sure scatter-gather list is large enough to accommodate all chunks
When decrypting a packet we arbitrary use 2 additional elements of the
sg list to allocate the packet header and the auth-tag (because in the
OpenVPN protocol the latter does not come at the end of the packet, but
rather between the header and the payload).

Ensure the sg list is large enough to accommodate all the elements.

Reported-by: Tony He <huangya90@gmail.com>
Reported-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>