ovpn-dco: netns-test.sh - automatically execute ping at the end
To make the test complete, automatically execute a ping from peer0
to peer1 after setup. This way it is possible to immediately have a
feedback on whether the tunnel is working or not.

Failing to ping the other peer will result in the netns-test.sh
script to exit with error.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>