ovpn-dco: restructure code to allow supporting multiple link to peers
This is an invasive change which refactors the way ovpn-dco handles
peers.
It has moved from storing a single peer to having a list of peers to
talk to.

Each peer is stored in at most 3 hashtables:
* by Peer ID (for fast lookup upon reception of DATA_V2 packets)
* by VPN IP (for fast lookup upon sending payload packets)
* by transport family/addr/port (for fast lookup upon reception of CTRL
  packets)

No iroute mechanism has been implemented, while ovpn-dco directly uses
the Linux routing table to figure out which nexthop should be used by an
outgoing packet. That next hop is then lookup in the VPN IP routing
table and a destination peer is retrieved.

This also means that packets having an unknown destination will be
discarded.

Due to the nature of the peer link separation, it is possible to
add peers using different address families and different transport
protocols at the same time. This means that the same instance could talk
to a peer via IPv4/TCP and another peer via IPv6/UDP.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>