ovpn-dco: don't have sockaddr_in6 point to struct sockaddr
'struct sockaddr' is not large enough to host a full 'struct
sockaddr_in6' object. Therefore we can't have the latter point
to an area allocated for the former.

In ovpn_peer_lookup_vpn_addr() get rid of the 'struct sockaddr'
generic object and directly use sockaddr_in{,6} as keys of the
hashtable.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>