ovpn-dco: ovpn-cli - allow adding peers in batch using the same UDP socket
For testing purposes it can be convenient to add a bunch of UDP peers,
all using the same socket.

For this reason a new command 'new_multi_peer' has been added to the
ovpn-cli tool. This command will read the peers from file add them all
in ovpn-dco using the same UDP socket.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>