ovpn-dco: consider overhead when initializing tunnel MTU
At the moment, when creating the tun interface, ovpn-dco does not
consider the encapsulation overhead for the MTU computation.

This results in a default MTU of 1500 which creates issues if the
underlying interface has an MTU of 1500 as well.

Since MSSFIX will not be implemented in ovpn-dco, it is necessary to
take appropriate measures to initialize the tun interface with a
safe MTU value.

Do so by subtracting the encapsulation overhead from the standard MTU
(ETH_DATA_LEN) when computing the interface MTU.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>