ovpn-dco: ensure NAPI is always released in case of peer add failure
ovpn_peer_new starts the NAPI subsystem for the new peer.
If then this peer cannot be added to the session (i.e. because the peer
ID already exists) the NAPI state should then be released.

Ensure the above by rearranging the cleanup peer routines so that NPI is
always reeased, also in case of an early failure.

This fixes an erroneous memory access upon interface deletion.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>