ovpn-dco: reject data packets from userspace
It is important to only have one place to generate packet IVs. Otherwise
a reuse of a IV could happen.

Reject data packet from user space as extra insurance against this.

This commit also adds a lower bound of 4 bytes to packets and
adds the peer-id to the debug log message.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>