ovpn-dco: avoid writing out of buffer when copying in6_addr object
When creating a family agnostic sockaddr, the appropriate type
"sockaddr_storage" should be used, because it is large enough to
accommodate both an IPv4 or IPv6 sockaddr.

Using "struct sockaddr" results in accessing memory out of boundary,
when treating it as a sockaddr_in6 (the latter is larger.

Fix out of bound memory access by using "sockaddr_storage" instead of
the smaller "sockaddr".

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>