ovpn-dco: do not return 0 when successfully processing TCP packet
The ovpn_recv() function returns 0 in case of success and this value is
propagated up to the caller of ovpn_tcp_rx_one().

This is a problem, because ovpn_tcp_rx_one() is expected to return the
amount of bytes read from the TCP socket. By returning 0, the caller
believes that there is nothing on the TCP stream and put the receiving
thread to sleep. A subsequent data_ready notification will then wake up
the receiving thread again.

This issue obviously triggers spurious delays when processing TCP data,
which results in stalls and bad performance.

Fix this by not overwriting the amount of bytes read from the socket
with the result of ovpn_recv().

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>