ovpn-dco: decrease reference counter of TCP socket instead of releasing
The TCP socket release path was accidentally calling sock_release(),
which would immediately take care of free'ing the socket, instead of
simply decreasing the reference counter.

ovpn-dco should not be responsible for releasing the socket, but it
should only take a reference and release it as needed.

The reference counting mechanism will take care of free'ing it at the
right time.

While at it, move the sockfd_put() in a common place to avoid having it
invoked once in the TCP and once in the UDP path.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>