ovpn-dco: tests - introduce multihome test script
With this script ovpn-dco is configured with multiple IPs
on an bridge interface called br0.
All vethX ifaces used to reach the various peers are enslaved in br0.

Each peer will reach the "server" (peer0) using a different IP.
The server is expected to reply using the same IP it received
data on.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>