ovpn-dco: add support for dumping peer status to userspace
With this patch a new netlink command is added: GET_PEER.
When invoked with the NLM_F_DUMP flag it will instruct the
ovpn-dco kernel module to send all known peers back to the
user.

At the moment only dumping is supported.
Single peer extraction will be implemented in a later patch.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>