ovpn-dco: tcp - get data pointer after having linearized skb
When linearizing a skb, the internal buffers may need to be reallocated.
When this happens the pointer to skb->data may change in order to point
to the new memory area.

For this reason, code should not rely on the value of skb->data
retrieved before invoking skb_linearize().

Fix ovpn_tcp_send_one() by retrieving the data pointer only after
having linearized the skb. This avoids bogus memory access with
consequent TCP connection stalls.

Reported-by: Lev Stipakov <lev@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>