ovpn-dco: ensure peers are indexed by transport address too
When adding a peer to ovpn-dco, the peer is added to a bunch of
hash tables for fast lookup.

Due to a mistake, the peer is being removed from the "has by transport
address" table right after being added, thus leaving this table always
empty.

Fix the issue by moving initialization at the beginning of the function.

Introduced-by: 80829387 ("ovpn-dco: avoid spurious warning from missing default case")
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>