ovpn-dco: use sockaddr_strorage to carry around family agnostic sockaddrs
When carrying around a pointer to a sockaddr which family is not yet
known, a sockaddr_storage pointer/object should always be used as its
size is large enough to contain any possible sockaddr. This way static
analyzer don't get fooled by what we are doing and won't throw any out
of bound error.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>