ovpn-dco: kill primary key if we ran out of IVs
With most ciphers we should never re-use the same IV more than once.
This is also bad practice in general.

For this reason, throw an error when we run out of IVs and delete the
key being used for encryption (located in the primary slot).

At this point no more packets can be encrypted, unless a new key is
installed.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>