ovpn-dco: require peer VPN IPs only in multipoint mode
In P2P mode there is no need to know the VPN IP of the peer on the other
end, because in any case, any traffic entering the tunnel will be sent
to that specific peer as there is no routing decision to make.

Therefore allow adding a peer without any VPN IP when in P2P mode.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>