ovpn-dco: don't use U16_MAX as packet max length in netlink policy
A netlink policy object stores the max and min value in s16 members.
This means that when using U16_MAX as value, the actual representation
become a negative number.

This condition triggers the warning at lib/nlattr.c:117, which checks
that min and max values are non-negative. Note that this is just a
cosmetic issue, because values are later copied into the
netlink_range_validation members which are u64 and thus properly
interpreted.

In any case, we don't truly need max to be U16_MAX, therefore reduce the
value to 4k in order to fit a s16, while still being reasonable for our
use case.

Fixes: https://github.com/OpenVPN/ovpn-dco/issues/21
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>