ovpn-dco: fix compat check for sock_is_readable on RHEL8
As for other parts of the code, RHEL8 ship an old kernel, but with a lot
of backported code from more recent kernels.

When our compat code tries to redefine some parts that fall into the
above, we will get an error because RHEL8 is fooling us regarding the
actual code version.

This is now happening for the definition of sock_is_readable.
Add extra EL8 check to make sure the compat code is not used for RHEL8
in this case.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>