ovpn-dco: Fix non-RHEL kernel builds
Red Hat Enterprise Linux kernels provides helper macros for
detecting the distribution version.  This is needed here as
Red Hat backports features and changes from newer kernels
into the older kernel baseline.  Therefore the RHEL kernel
features may not be correctly identified by the Linux kernel
version alone.

To be able to build ovpn-dco on non-RHEL kernels, we need
these helper macros defined.  And we want the result to
always be true, to not disable the other kernel version
checks

Signed-off-by: David Sommerseth <davids@openvpn.net>
Signed-off-by: Antonio Quartulli <antonio@openvpn.net>